Desktop/My prorfiloi/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
Jinja2==3.1.2
Flask-Admin==1.6.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
Flask-Mail==0.9.1
WTForms==3.1.1
email-validator==2.1.0.post1
Flask-Compress==1.14
gunicorn==21.2.0